+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import universal_navigation
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  UniversalNavigationPlugin.register(with: registry.registrar(forPlugin: "UniversalNavigationPlugin"))
+}