+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+#include "generated_plugin_registrant.h"
+
+#include <universal_navigation/universal_navigation_plugin.h>
+
+void RegisterPlugins(flutter::PluginRegistry* registry) {
+  UniversalNavigationPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("UniversalNavigationPlugin"));
+}