+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.5.0
 
 * Added documentation of **AppNavigator**,**CoreEventNotifier** and **u_nav_configurator.dart** .
-* Added supprot macos, windows and web.
+* Added support macos, windows and web.
 
 ## 0.4.4